Documented internal ShaEngineHandler() function calls
Documented what should be all of the internal function calls within the HMAC portion of the SHA engine.

There will need to be some RE done in findKeyTypeUnmasked() to find out how the buffer keyArray[] works. I think this isn't directly the key's data but maybe some sort of struct that acts as a metadata describing the key's attributes?