chore: define a type for thread functions
Having the main() thread function decelared as a u32 forces the caller
to write a cast. Let's define a type for thread functions: this not only
can help to reduce casts, but also serves as documentation on what
signature thread functions should have.

We use `int` instead of `s32` (or, worse, `u32`) since that's a bit more
natural (while being effectively the same thing).