chore: change thread priority to be signed
OSGetThreadPriority returns a signed value, since it uses negative
values for errors. Since the priority cannot be negative anyway, let's
use a signed type.

This is also more consistent with the original kernel, that is using the
ARM BLT instruction to decide which branch to take after a priority
comparison (BLT operates on signed registers).

This also lets us remove some casts.